requirements.txt
google_api_python_client==2.74.0
google_auth_oauthlib==0.4.4
pandas==1.3.5
protobuf==4.21.12
